requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
alembic==1.16.2
ane_transformers==0.1.1
antlr4-python3-runtime==4.9.3
asteroid-filterbanks==0.4.0
attrs==25.3.0
av==14.4.0
cattrs==25.1.1
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.
coloredlogs==15.0.1
colorlog==6.9.0
contourpy==1.3.2
coremltools==8.3.0
ctranslate2==4.4.0
cycler==0.12.1
docopt==0.6.2
einops==0.8.1
faster-whisper==1.1.1
ffmpeg-python==0.2.0
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.58.4
frozenlist==1.7.0
fsspec==2025.5.1
future==1.0.0
hf-xet==1.1.4
hf_transfer==0.1.9
huggingface-hub==0.33.0
humanfriendly==10.0
HyperPyYAML==1.2.2
idna==3.10
inquirerpy==0.3.4
Jinja2==3.1.6
joblib==1.5.1
julius==0.2.7
kiwisolver==1.4.8
lightning==2.5.1.post0
lightning-utilities==0.14.3
llvmlite==0.44.0
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
more-itertools==10.7.0
mpmath==1.3.0
multidict==6.4.4
networkx==3.5
nltk==3.9.1
numba==0.61.2
numpy==2.2.6
omegaconf==2.3.0
onnxruntime==1.22.0
openai-whisper==20240930
optuna==4.4.0
packaging==24.2
pandas==2.3.0
pfzy==0.3.4
pillow==11.2.1
primePy==1.3
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==3.20.1
psutil==7.0.0
pyaml==25.5.0
pyannote.audio==3.3.2
pyannote.core==5.0.0
pyannote.database==5.1.3
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pycparser==2.22
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytorch-lightning==2.5.1.post0
pytorch-metric-learning==2.8.1
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
rich==14.0.0
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.15.3
semver==3.0.4
sentencepiece==0.2.0
shellingham==1.5.4
silero-vad==5.1.2
six==1.17.0
sortedcontainers==2.4.0
soundfile==0.13.1
speechbrain==1.0.3
SQLAlchemy==2.0.41
sympy==1.14.0
tabulate==0.9.0
tensorboardX==2.6.4
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
torch==2.7.1
torch-audiomentations==0.12.0
torch_pitch_shift==1.2.5
torchaudio==2.7.1
torchmetrics==1.7.3
tqdm==4.67.1
transformers==4.52.4
typer==0.16.0
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
wcwidth==0.2.13
whisperx==3.3.4
yarl==1.20.1
